Fix crash when connecting links
Prior to ComfyUI_frontend 1.4.3, there was an edge case where both connecting links `input` and `output` could be null.

As of 1.4.3, this is always the case when a user clicks a slot or reroute, but does not drag it anywhere.  The expected UX is "cancel, no changes made".

Hope this helps.